+++ CHANGELOG.md
@@ -1,11 +1,12 @@
 Change Log
 ===
 
-* [feature] added `no-unused-expression` rule which disallows unused expression statements
+* [new-rule] added `no-unused-expression` rule which disallows unused expression statements
+* [feature] the `check-operator` option for the `whitespace` rule now checks whitespace around the => token
+* [bug] `no-use-before-declare-rule` no longer triggers false positives for member variables of classes used before the class is declared
 * [bug] semicolon at end of file no longer triggers false positives for `whitespace` rule
 * [bug] hoisted functions no longer cause false positives for the `no-unreachable` rule
 * [bug] the rule loader no longer transforms/ignores the leading and trailing underscores and dashes of rule names in the config file
-* [feature] the `check-operator` option for the `whitespace` rule now checks whitespace around the => token
 * [bug] `export import` statements no longer false positives for `no-unused-variable-rule`
 
 v0.4.6